Extract ConnectorImportOperations from AppsPage into its own file
